# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CloudFuse API",
    "description": "Multi-language PaaS Platform API documentation",
    "version": "1.0",
    "contact": {
      "name": "CloudFuse Team",
      "email": "support@cloudfuse.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.cloudfuse.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/deployments": {
      "get": {
        "tags": [
          "Deployments"
        ],
        "summary": "Get all deployments",
        "description": "Retrieve all deployments for the authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of deployments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Deployment"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 3
                    },
                    "page": {
                      "type": "number",
                      "example": 1
                    },
                    "limit": {
                      "type": "number",
                      "example": 10
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Deployments"
        ],
        "summary": "Create new deployment",
        "description": "Create a new deployment for an application",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeploymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Deployment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deployment"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication"
      }
    },
    "schemas": {
...
      "Deployment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "building",
              "active",
              "failed"
            ],
            "example": "active"
          },
          "url": {
            "type": "string",
            "example": "https://my-app.cloudfuse.com"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateDeploymentDto": {
        "type": "object",
        "properties": {
          "appId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "version": {
            "type": "string",
            "example": "1.0.0"
          },
          "environment": {
            "type": "string",
            "example": "production"
          }
        },
        "required": [
          "appId"
        ]
      },
...
    }
  },
...
}